Finish reimpl. of grid state + win tracking
GridState class is now plugged into the game state. GridState is not
used in history directly but instead a mirror data structure is
stored in history. Grid is based on a single GridState object that is
updated based upon current step.

Win checking uses adjacent sequence finder. This now opens up the
possibilities for larger boards and other win conditions.